Move api/router/util to util/router (#1909)
