+++ api/router/registry/registry.go
@@ -12,11 +12,11 @@ import (
 
 	"github.com/micro/go-micro/v3/api"
 	"github.com/micro/go-micro/v3/api/router"
-	"github.com/micro/go-micro/v3/api/router/util"
 	"github.com/micro/go-micro/v3/logger"
 	"github.com/micro/go-micro/v3/metadata"
 	"github.com/micro/go-micro/v3/registry"
 	"github.com/micro/go-micro/v3/registry/cache"
+	util "github.com/micro/go-micro/v3/util/router"
 )
 
 // endpoint struct, that holds compiled pcre
