+++ api/router/static/static.go
@@ -10,11 +10,11 @@ import (
 
 	"github.com/micro/go-micro/v3/api"
 	"github.com/micro/go-micro/v3/api/router"
-	"github.com/micro/go-micro/v3/api/router/util"
 	"github.com/micro/go-micro/v3/logger"
 	"github.com/micro/go-micro/v3/metadata"
 	"github.com/micro/go-micro/v3/registry"
 	rutil "github.com/micro/go-micro/v3/util/registry"
+	util "github.com/micro/go-micro/v3/util/router"
 )
 
 type endpoint struct {
