+++ util/router/types_test.go
@@ -1,4 +1,4 @@
-package util
+package router
 
 // download from https://raw.githubusercontent.com/grpc-ecosystem/grpc-gateway/master/protoc-gen-grpc-gateway/httprule/types_test.go
